Création d'un objet spécifique pour les départements

Cette fonction permet de créer un objet de type departement à partir d'un dataframe contenant des informations sur les communes, les élus et les collectivités. Elle extrait des informations essentielles concernant un département y compris les communes qui le composent, les élus et leurs âges, ainsi que des informations supplémentaires sur les collectivités à statut particulier.

Usage

creer_departement(df)

Arguments

df

Un dataframe contenant les informations suivantes :

Code.du.département : Code du département.

Libellé.du.département : Nom du département.

Code.de.la.commune : Code de la commune.

Libellé.de.la.commune : Nom de la commune.

Code.de.la.collectivité.à.statut.particulier (optionnel) : Code de la collectivité à statut particulier.

Libellé.de.la.collectivité.à.statut.particulier (optionnel) : Nom de la collectivité à statut particulier.

Value

Un objet de type departement, qui est une liste contenant les informations suivantes :

Code.du.département : Code unique du département.

Libellé.du.département : Nom du département.

Collectivite : Liste des collectivités à statut particulier si elles existent dans les données (sinon NULL).

Communes : Liste des objets de type commune créés pour chaque commune présente dans le département.